// clients/js/src/generated/accounts/coachClientCredits.ts
export type CoachClientCredits = {
  discriminator: ReadonlyUint8Array;
  version: number;
  coachAuthority: Address;
  clientWallet: Address;
  availableCredits: bigint;
  reservedCredits: bigint;
  totalPurchased: bigint;
  purchaseCount: bigint;
  nextPurchaseNonce: bigint;
  lastOffer: Address;
  lastPurchaseAt: bigint;
  bump: number;
  reserved: Array<number>;
};