// clients/js/src/generated/types/creditReservationStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum CreditReservationStatus {
  Reserved,
  Returned,
  Consumed,
}

export type CreditReservationStatusArgs = CreditReservationStatus;

export function getCreditReservationStatusEncoder(): FixedSizeEncoder<CreditReservationStatusArgs> {
  return getEnumEncoder(CreditReservationStatus);
}

export function getCreditReservationStatusDecoder(): FixedSizeDecoder<CreditReservationStatus> {
  return getEnumDecoder(CreditReservationStatus);
}

export function getCreditReservationStatusCodec(): FixedSizeCodec<
  CreditReservationStatusArgs,
  CreditReservationStatus
> {
  return combineCodec(
    getCreditReservationStatusEncoder(),
    getCreditReservationStatusDecoder(),
  );
}